requirements.txt
apache-airflow-providers-mongo
airflow-clickhouse-plugin
pymongo
psycopg2-binary
pandas
boto3
python-telegram-bot
